requirements.txt
streamlit==1.32.0
pytube==15.0.0
openai-whisper==20231106
transformers==4.37.0
nltk==3.8.1
gitpython==3.1.40
